+++ src/debug.c
@@ -1781,7 +1781,7 @@ static void stub_test_task(void* arg)
         TEST_TRY_FUNC(task_create("test", 0x1c, 0x1000, test_task, 0));
         msleep(100);
         TEST_TRY_FUNC_CHECK(test_task_created, == 1);
-        TEST_TRY_FUNC_CHECK_STR(get_task_name_from_id((unsigned int)get_current_task()), "run_test");
+        TEST_TRY_FUNC_CHECK_STR(get_task_name_from_id(get_current_task()), "run_test");
 
         // mq
         static struct msg_queue * mq = 0;
